[Smoke Tests] Split up massive smoke-tests.rs file into components.
