+++ testsuite/smoke-test/src/lib.rs
@@ -2,7 +2,40 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #[cfg(test)]
-mod smoke_test;
+mod client;
+
+#[cfg(test)]
+mod full_nodes;
+
+#[cfg(test)]
+mod genesis;
+
+#[cfg(test)]
+mod key_manager;
+
+#[cfg(test)]
+mod operational_tooling;
+
+#[cfg(test)]
+mod replay_tooling;
+
+#[cfg(test)]
+mod scripts_and_modules;
+
+#[cfg(test)]
+mod state_sync;
+
+#[cfg(test)]
+mod storage;
+
+#[cfg(test)]
+mod test_environment;
+
+#[cfg(test)]
+mod test_utils;
+
+#[cfg(test)]
+mod transaction;
 
 #[cfg(test)]
 mod workspace_builder;
